requirements.txt
psycopg2-binary
sqlalchemy
pydantic
fastapi
uvicorn
python-dotenv
filestack-python
openai